Add previous check check to production check
This PR should change our slack notify job to only alert on a production check failure if the previous run has also failed.

We do see one off failures of the job, but more often than not these are isolated hiccups that we don't need to worry about.  This PR ensures that if there is a persistent failure we get the notification rather than suffering from notification burnout and ignoring.